LPS-94301 wiki - classes in module that are not exported should be in internal package
