+++ modules/apps/wiki/wiki-web/src/main/java/com/liferay/wiki/web/internal/display/context/util/WikiRequestHelper.java
@@ -27,7 +27,7 @@ import com.liferay.wiki.configuration.WikiGroupServiceOverriddenConfiguration;
 import com.liferay.wiki.constants.WikiConstants;
 import com.liferay.wiki.constants.WikiWebKeys;
 import com.liferay.wiki.model.WikiPage;
-import com.liferay.wiki.web.configuration.WikiPortletInstanceConfiguration;
+import com.liferay.wiki.web.internal.configuration.WikiPortletInstanceConfiguration;
 
 import javax.servlet.http.HttpServletRequest;
