+++ modules/apps/wiki/wiki-web/src/main/java/com/liferay/wiki/web/internal/portlet/action/ActionUtil.java
@@ -44,7 +44,7 @@ import com.liferay.wiki.service.WikiNodeLocalServiceUtil;
 import com.liferay.wiki.service.WikiNodeServiceUtil;
 import com.liferay.wiki.service.WikiPageLocalServiceUtil;
 import com.liferay.wiki.service.WikiPageServiceUtil;
-import com.liferay.wiki.web.configuration.WikiPortletInstanceConfiguration;
+import com.liferay.wiki.web.internal.configuration.WikiPortletInstanceConfiguration;
 import com.liferay.wiki.web.internal.security.permission.resource.WikiNodePermission;
 import com.liferay.wiki.web.internal.util.WikiUtil;
 import com.liferay.wiki.web.internal.util.WikiWebComponentProvider;
