+++ modules/apps/wiki/wiki-web/src/main/resources/META-INF/resources/wiki/init.jsp
@@ -42,7 +42,7 @@ page import="com.liferay.wiki.exception.WikiFormatException" %><%@
 page import="com.liferay.wiki.social.WikiActivityKeys" %><%@
 page import="com.liferay.wiki.util.comparator.PageVersionComparator" %><%@
 page import="com.liferay.wiki.validator.WikiPageTitleValidator" %><%@
-page import="com.liferay.wiki.web.configuration.WikiPortletInstanceConfiguration" %><%@
+page import="com.liferay.wiki.web.internal.configuration.WikiPortletInstanceConfiguration" %><%@
 page import="com.liferay.wiki.web.internal.display.context.WikiDisplayContextProvider" %><%@
 page import="com.liferay.wiki.web.internal.display.context.logic.MailTemplatesHelper" %><%@
 page import="com.liferay.wiki.web.internal.display.context.logic.WikiPortletInstanceSettingsHelper" %><%@
